fix(module:message): fix lazy load problem (#3797)
* fix(module:message): fix lazy load problem

close #3794

* test(module:message): fix test

* fix(module:message): fix module
