+++ components/message/nz-message.module.ts
@@ -15,9 +15,10 @@ import { NzIconModule } from 'ng-zorro-antd/icon';
 import { NZ_MESSAGE_DEFAULT_CONFIG_PROVIDER } from './nz-message-config';
 import { NzMessageContainerComponent } from './nz-message-container.component';
 import { NzMessageComponent } from './nz-message.component';
+import { NzMessageServiceModule } from './nz-message.service.module';
 
 @NgModule({
-  imports: [CommonModule, OverlayModule, NzIconModule, NzAddOnModule],
+  imports: [CommonModule, OverlayModule, NzIconModule, NzAddOnModule, NzMessageServiceModule],
   declarations: [NzMessageContainerComponent, NzMessageComponent],
   providers: [NZ_MESSAGE_DEFAULT_CONFIG_PROVIDER],
   entryComponents: [NzMessageContainerComponent]
