+++ components/notification/nz-notification.service.ts
@@ -14,9 +14,10 @@ import { NzMessageBaseService } from 'ng-zorro-antd/message';
 import { NzNotificationConfig } from './nz-notification-config';
 import { NzNotificationContainerComponent } from './nz-notification-container.component';
 import { NzNotificationData, NzNotificationDataFilled, NzNotificationDataOptions } from './nz-notification.definitions';
+import { NzNotificationServiceModule } from './nz-notification.service.module';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: NzNotificationServiceModule
 })
 export class NzNotificationService extends NzMessageBaseService<
   NzNotificationContainerComponent,
