+++ components/notification/public-api.ts
@@ -11,4 +11,5 @@ export * from './nz-notification.component';
 export * from './nz-notification.module';
 export * from './nz-notification.definitions';
 export * from './nz-notification.service';
+export * from './nz-notification.service.module';
 export * from './nz-notification-container.component';
